+++ changelog.md
@@ -28,6 +28,7 @@ Last change: 26-JUN-2026 CEST
 - Improved measure tool
 - Faster AIM through smarter implementation
 - Faster and more memory efficient (especially for large datasets) SMLM clusterer + progress bar added
+- Progress bar for finding cluster centers
 
 #### SPINNA
 - Fixed verbose for batch analysis